Add more methods Optimise tcpStreamReadUntilSpace

+++ src/tcp_stream.h
@@ -13,9 +13,9 @@
 typedef struct TcpStream {
     int fd;
     int error;
-    int cursor;
-    int length;
-    int capacity;
+    size_t cursor;
+    size_t length;
+    size_t capacity;
     char *buffer;
 } TcpStream;
 
@@ -26,23 +26,23 @@ void freeTcpStream(TcpStream *stream);
 /* Waits for data to be available. */
 void tcpStreamWait(TcpStream *stream);
 /* Fills the internal buffer at least length. */
-void tcpStreamFill(TcpStream *stream, int length);
+void tcpStreamFill(TcpStream *stream, size_t length);
 /* Advances the cursor by size. Writes in the buffer. */
-void tcpStreamRead(TcpStream *stream, void *buffer, int size);
+void tcpStreamRead(TcpStream *stream, void *buffer, size_t size);
 /* Advances the cursor by size. Returns ptr. */
-void *tcpStreamReadSlice(TcpStream *stream, int size);
+void *tcpStreamReadSlice(TcpStream *stream, size_t size);
 /* Drains internal buffer until cursor. Performs memmove. */
 void tcpStreamDrain(TcpStream *stream);
 /* Reads and returns one byte at cursor, moves the cursor. */
 char tcpStreamGetc(TcpStream *stream);
 /* Reads and returns one byte at index. */
-char tcpStreamIndex(TcpStream *stream, int index);
+char tcpStreamIndex(TcpStream *stream, size_t index);
 /* Reads until space character. */
-string tcpStreamReadUntilSpace(TcpStream *stream, int maxLength);
+string tcpStreamReadUntilSpace(TcpStream *stream, size_t maxLength);
 /* Read until carriage return and new line. */
-string tcpStreamReadUntilCRLF(TcpStream *stream, int maxLength, int ignoreLoneCRLF);
+string tcpStreamReadUntilCRLF(TcpStream *stream, size_t maxLength, int ignoreLoneCRLF);
 /* Read until string, returns the slice with length not including the substr. */
-string tcpStreamReadUntilString(TcpStream *stream, int maxLength, const char *str, int size);
+string tcpStreamReadUntilString(TcpStream *stream, size_t maxLength, const char *str, size_t size);
 int tcpStreamHasError(TcpStream *stream);
 
 #endif //TCP_STREAM_H
